requirements.txt
openai
python-dotenv
pydantic
colorama
matplotlib
seaborn
pandas
